# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MovieFlix API",
    "version": "1.0.0",
    "description": "API responsável por gerenciar os filmes do MovieFlix"
  },
  "paths": {
    "/movies": {
      "get": {
        "summary": "Listar todos os filmes",
        "tags": ["Movies"],
        "responses": {
          "200": {
            "description": "Lista de filmes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar novo filme",
        "tags": ["Movies"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Filme criado com sucesso"
          },
          "409": {
            "description": "Filme com mesmo título já existe"
          }
        }
      }
    },
    "/movies/{id}": {
      "put": {
        "summary": "Atualizar um filme",
        "tags": ["Movies"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Filme atualizado com sucesso"
          },
          "404": {
            "description": "Filme não encontrado"
          }
        }
      },
      "delete": {
        "summary": "Deletar um filme",
        "tags": ["Movies"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filme deletado com sucesso"
          },
          "404": {
            "description": "Filme não encontrado"
          }
        }
      }
    },
    "/movies/{genreName}": {
      "get": {
        "summary": "Filtrar filmes por nome do gênero",
        "tags": ["Movies"],
        "parameters": [
          {
            "name": "genreName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de filmes filtrada por gênero",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "genre_id": {
            "type": "integer"
          },
          "language_id": {
            "type": "integer"
          },
          "oscar_count": {
            "type": "integer"
          },
          "release_date": {
            "type": "string",
            "format": "date"
          },
          "genres": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "name": { "type": "string" }
            }
          },
          "languages": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "name": { "type": "string" }
            }
          }
        }
      },
      "MovieInput": {
        "type": "object",
        "required": ["title", "genre_id", "language_id", "oscar_count", "release_date"],
        "properties": {
          "title": {
            "type": "string"
          },
          "genre_id": {
            "type": "integer"
          },
          "language_id": {
            "type": "integer"
          },
          "oscar_count": {
            "type": "integer"
          },
          "release_date": {
            "type": "string",
            "format": "date"
          }
        }
      }
    }
  }
}